add forward/all view (maybe should rename new.js?)
